Verificacao se comandos ja existem antes da instalacao

+++ grdk-install-worker.sh
@@ -5,17 +5,30 @@ set -e
 apt-get update && apt-get install -y nfs-common unzip libconfig-yaml-perl libjson-xs-perl jq
 
 ### DOCKER-CE
-curl -sSL https://get.docker.com | CHANNEL=stable sh
-echo "{\"dns\":[\"${DK_SERVER_DNS}\"],\"insecure-registries\":[\"${DK_REPO_DI_HOST}:5000\"]}" > /etc/docker/daemon.json
+if command_exists docker; then
+	echo "Docker already installed"
+	echo $(docker -v | cut -d ' ' -f3 | cut -d ',' -f1)
+else
+	curl -sSL https://get.docker.com | VERSION=18.06.3-ce sh
+	echo "{\"dns\":[\"${DK_SERVER_DNS}\"],\"insecure-registries\":[\"${DK_REPO_DI_HOST}:5000\"]}" > /etc/docker/daemon.json
+fi
 
 ### DOCKER-COMPOSE
-curl -L https://github.com/docker/compose/releases/download/1.18.0/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
-chmod +x /usr/local/bin/docker-compose
+if command_exists docker-compose; then
+	echo "Docker Composer already installed"
+else
+	curl -L https://github.com/docker/compose/releases/download/1.18.0/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
+	chmod +x /usr/local/bin/docker-compose
+fi
 
 ### DOCKER-VOLUME-NETSHARE
-wget https://github.com/ContainX/docker-volume-netshare/releases/download/v0.34/docker-volume-netshare_0.34_amd64.deb
-dpkg -i docker-volume-netshare_0.34_amd64.deb
-rm docker-volume-netshare_0.34_amd64.deb
+if command_exists docker-volume-netshare; then
+	echo "Docker Volume Netshare already installed"
+else
+	wget https://github.com/ContainX/docker-volume-netshare/releases/download/v0.34/docker-volume-netshare_0.34_amd64.deb
+	dpkg -i docker-volume-netshare_0.34_amd64.deb
+	rm docker-volume-netshare_0.34_amd64.deb
+fi
 
 ### SERVICES UP
 if [[ ! `ps ax | grep dockerd | grep -v grep` ]]; then
